fix: 카테고리 카드를 버튼으로 변경
